backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
python-dotenv==1.1.0
supabase==2.15.2
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
